Allow multidex of instrumentations
Allow installation of instrumentation secondary dex files. Both
the instrumentation and the instrumented application are installed (if
necessary) by MultiDex.installInstrumentation. Instrumentation
secondary dex files are extracted in the Application code cache folder
because it generally doesn't have access to its own folder.
Instrumentation preferences are saved in the Application preferences
with a prefix.

Bug: 31383194

Test: frameworks/base/core/tests/hosttests/test-apps/MultiDexLegacy*

Change-Id: I705ed87162326fd64128454aa144a359b09436cd